+++ changelog.md
@@ -9,6 +9,7 @@
 
 - Added support for PlayerRoles Syncing (https://modrinth.com/mod/player-roles)
 - Added support for CobblemonGuilds
+- Allow disabling verification via DMs
 
 **Changes**: